+++ core/scripts/migration/mysql_6_to_4.sql
@@ -1,0 +1,3 @@
+alter table Tables drop column owner_table;
+alter table Tables drop column partition_tag;
+alter table Tables drop column version;