+++ core/src/sdk/include/MilvusApi.h
@@ -81,11 +81,11 @@ struct RowRecord {
 /**
  * @brief TopK query result
  */
-struct TopKQueryResult {
-    int64_t row_num;
+struct QueryResult {
     std::vector<int64_t> ids;
     std::vector<float> distances;
 };
+using TopKQueryResult = std::vector<QueryResult>;
 
 /**
  * @brief index parameters